Remove freetype extension
We needed this extension on zLinux because the absence of the
pkg-config package causes us to go down strange code paths. As
long as we mandate that package (already in dockerfiles and soon
to be in the AoK Ansible playbooks) is used for builds, we should
have no need for this extension.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>